Flytt data som kun trengs i modalen inn i modal-komponent
